feat(fuji): add URL-based entry routing with /entries/[id]
Entry IDs now appear in the URL when editing. Clicking an entry
navigates to /entries/{id} instead of toggling in-memory state.

- Move app shell (header, sidebar, resizable panes, command palette)
  from +page.svelte into +layout.svelte so it persists across routes
- Create /entries/[id] route that reads page.params.id and opens
  the CRDT document handle
- Strip +page.svelte to list-only view (table/timeline)
- Replace all viewState.selectEntry() calls with goto() navigation
- Entry URLs are bookmarkable, shareable, and survive page refresh
