+++ apps/fuji/src/routes/+page.svelte
@@ -1,53 +1,11 @@
 <script lang="ts">
 	import { Button } from '@epicenter/ui/button';
-	import {
-		CommandPalette,
-		type CommandPaletteItem,
-	} from '@epicenter/ui/command-palette';
-	import * as Resizable from '@epicenter/ui/resizable';
-	import type { DocumentHandle } from '@epicenter/workspace';
 	import ClockIcon from '@lucide/svelte/icons/clock';
-	import FileTextIcon from '@lucide/svelte/icons/file-text';
 	import TableIcon from '@lucide/svelte/icons/table-2';
-	import type * as Y from 'yjs';
-	import { workspace } from '$lib/client';
-	import AppHeader from '$lib/components/AppHeader.svelte';
 	import EntriesTable from '$lib/components/EntriesTable.svelte';
-	import EntryEditor from '$lib/components/EntryEditor.svelte';
 	import EntryTimeline from '$lib/components/EntryTimeline.svelte';
-	import EntriesSidebar from '$lib/components/EntriesSidebar.svelte';
-	import { Kbd } from '@epicenter/ui/kbd';
 	import { entriesState } from '$lib/entries.svelte';
 	import { viewState } from '$lib/view.svelte';
-
-	// ─── Command Palette ─────────────────────────────────────────────────────────
-
-	let paletteOpen = $state(false);
-	let paletteQuery = $state('');
-
-	const paletteItems = $derived.by((): CommandPaletteItem[] => {
-		if (!paletteOpen) return [];
-		return entriesState.active.map((entry) => ({
-			id: entry.id,
-			label: entry.title || 'Untitled',
-			description: entry.subtitle || undefined,
-			icon: FileTextIcon,
-			keywords: [...entry.tags, ...entry.type],
-			group: entry.type.length > 0 ? entry.type[0] : 'Uncategorized',
-			onSelect: () => viewState.selectEntry(entry.id),
-		}));
-	});
-
-	// ─── Document Handle (Y.XmlFragment) ────────────────────────────────────────────
-
-	let currentYXmlFragment = $state<Y.XmlFragment | null>(null);
-	let currentDocHandle = $state<DocumentHandle | null>(null);
-
-	const selectedEntry = $derived(
-		viewState.selectedEntryId
-			? (entriesState.get(viewState.selectedEntryId) ?? null)
-			: null,
-	);
 
 	/** Entries filtered by sidebar type/tag filters. */
 	const filteredEntries = $derived.by(() => {
@@ -62,118 +20,29 @@
 		}
 		return result;
 	});
-
-
-	$effect(() => {
-		const entryId = viewState.selectedEntryId;
-		if (!entryId) {
-			currentYXmlFragment = null;
-			currentDocHandle = null;
-			return;
-		}
-
-		let cancelled = false;
-		workspace.documents.entries.content.open(entryId).then((handle) => {
-			if (cancelled) return;
-			currentDocHandle = handle;
-			currentYXmlFragment = handle.asRichText();
-		});
-
-		return () => {
-			cancelled = true;
-			if (currentDocHandle) {
-				workspace.documents.entries.content.close(entryId);
-			}
-			currentYXmlFragment = null;
-			currentDocHandle = null;
-		};
-	});
 </script>
 
-<svelte:window
-	onkeydown={(event) => {
-	const isInputFocused =
-		event.target instanceof HTMLInputElement ||
-		event.target instanceof HTMLTextAreaElement ||
-		(event.target instanceof HTMLElement && event.target.isContentEditable);
-
-
-	if (event.key === 'n' && event.metaKey) {
-		event.preventDefault();
-		entriesState.createEntry();
-		return;
-	}
-
-	if (event.key === 'Escape' && !isInputFocused && viewState.selectedEntryId) {
-		event.preventDefault();
-		viewState.selectEntry(null);
-	}
-}}
-/>
-
-<div class="flex h-screen flex-col">
-	<AppHeader onOpenSearch={() => (paletteOpen = true)} />
-	<Resizable.PaneGroup direction="horizontal" class="flex-1">
-		<Resizable.Pane defaultSize={20} minSize={15} maxSize={40}>
-			<EntriesSidebar entries={entriesState.active} />
-		</Resizable.Pane>
-		<Resizable.Handle withHandle />
-		<Resizable.Pane defaultSize={80}>
-			<main class="flex h-full flex-1 flex-col overflow-hidden">
-				{#if selectedEntry && currentYXmlFragment}
-					{#key viewState.selectedEntryId}
-						<EntryEditor
-							entry={selectedEntry}
-							yxmlfragment={currentYXmlFragment}
-						/>
-					{/key}
-				{:else if selectedEntry}
-					<div class="flex h-full items-center justify-center">
-						<p class="text-muted-foreground">Loading editor…</p>
-					</div>
-				{:else}
-					<!-- View mode toggle header -->
-					<div class="flex items-center justify-end border-b px-4 py-2">
-						<Button
-							variant="ghost"
-							size="icon"
-							class="size-7"
-							onclick={() => viewState.toggleViewMode()}
-							title={viewState.viewMode === 'table' ? 'Switch to timeline' : 'Switch to table'}
-						>
-							{#if viewState.viewMode === 'table'}
-								<ClockIcon class="size-4" />
-							{:else}
-								<TableIcon class="size-4" />
-							{/if}
-						</Button>
-					</div>
-
-					{#if viewState.viewMode === 'table'}
-						<EntriesTable entries={filteredEntries} />
-					{:else}
-						<EntryTimeline entries={filteredEntries} />
-					{/if}
-				{/if}
-			</main>
-		</Resizable.Pane>
-	</Resizable.PaneGroup>
-	<div class="flex h-6 shrink-0 items-center gap-3 border-t bg-background px-3 text-xs text-muted-foreground">
-		<span>{entriesState.active.length} {entriesState.active.length === 1 ? 'entry' : 'entries'}</span>
-		<div class="ml-auto flex items-center gap-1.5">
-			<span class="flex items-center gap-1">
-				Search <Kbd>⌘K</Kbd>
-			</span>
-		</div>
+<main class="flex h-full flex-1 flex-col overflow-hidden">
+	<!-- View mode toggle header -->
+	<div class="flex items-center justify-end border-b px-4 py-2">
+		<Button
+			variant="ghost"
+			size="icon"
+			class="size-7"
+			onclick={() => viewState.toggleViewMode()}
+			title={viewState.viewMode === 'table' ? 'Switch to timeline' : 'Switch to table'}
+		>
+			{#if viewState.viewMode === 'table'}
+				<ClockIcon class="size-4" />
+			{:else}
+				<TableIcon class="size-4" />
+			{/if}
+		</Button>
 	</div>
-</div>
 
-<CommandPalette
-	items={paletteItems}
-	bind:open={paletteOpen}
-	bind:value={paletteQuery}
-	placeholder="Search entries…"
-	emptyMessage="No entries found."
-	title="Search Entries"
-	description="Search entries by title, subtitle, tags, or type"
-/>
+	{#if viewState.viewMode === 'table'}
+		<EntriesTable entries={filteredEntries} />
+	{:else}
+		<EntryTimeline entries={filteredEntries} />
+	{/if}
+</main>
